feat: add cache header file
