+++ src/cache.c
@@ -1,3 +1,5 @@
+#include "cache.h"
+
 #include <errno.h>
 #include <fcntl.h>
 #include <stdio.h>
@@ -9,20 +11,6 @@
 
 #define BLOCK_SIZE 4096
 #define CACHE_SIZE 8
-
-typedef struct CacheBlock {
-  int fd;
-  off_t offset;
-  char data[BLOCK_SIZE];
-  struct CacheBlock *prev;
-  struct CacheBlock *next;
-} CacheBlock;
-
-typedef struct {
-  CacheBlock *head; // points to the oldest block
-  CacheBlock *tail; // points to the most recently added block
-  int current_size;
-} Cache;
 
 Cache cache = {NULL, NULL, 0}; // initialize empty cache
